Separation of tests into a dedicated file.
